requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
joblib==1.3.2
mysqlclient==2.2.0
networkx==3.1
numpy==1.26.0
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.3
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
